tetragon: Wait for go routines to complete for exit
When we receive a signal -- sigterm for example -- we tear down the BPF
state and event handlers in go routines with cancel(). But, we do not
wait to ensure they have received the signal and done any cleanup needed.

This adds a waitgroup to use for this. So objects increment it on New()
create op and then call the Done() when they are complete. Ensuring we
don't exit in the middle of sensitive operations, such as writes to gRPC
streams and JSON files where we can generate corrupted events when we
are unlucky.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>